requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gmpy2==2.2.1
gmpy2-stubs==2.2.1.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycryptodome==3.22.0
Werkzeug==3.1.3
